rename plp banner
