+++ packages/extension-chakra-storefront/src/pages/product-list/index.jsx
@@ -36,7 +36,7 @@ import SelectedRefinements from '../../pages/product-list/partials/selected-refi
 import EmptySearchResults from '../../pages/product-list/partials/empty-results'
 import PageHeader from './partials/product-list-title'
 import AbovePageHeader from '../../pages/product-list/partials/above-page-header'
-import PageDesignerPromotionalBanner from '../../pages/product-list/partials/page-designer-promotional-banner'
+import PageDesignerPromotionalBanner from './partials/product-list-banner'
 import PageMetadata from './page-metadata'
 import PageCache from './page-cache'
 import Sort from './partials/sort'
