+++ packages/extension-chakra-storefront/src/pages/product-list/partials/product-list-banner.jsx
@@ -13,11 +13,11 @@ import {Page} from '@salesforce/commerce-sdk-react/components'
 
 // Page Designer Components
 import {ImageWithText} from '../../../page-designer/assets'
-import {MobileGrid1r1c} from '../../..//page-designer/layouts'
+import {MobileGrid1r1c} from '../../../page-designer/layouts'
 
 // Hooks
 import {isServer} from '../../../utils/utils'
-import useMultiSite from '../../..//hooks/use-multi-site'
+import useMultiSite from '../../../hooks/use-multi-site'
 
 // Constants
 const PROMO_BANNER_DESKTOP_PAGE_ID = 'instagram-promo-banner-desktop'
